refactor: 💡 errorStatus, successStatus constant로 묶음
