+++ src/main/java/org/runnect/server/common/exception/BasicException.java
@@ -1,6 +1,7 @@
 package org.runnect.server.common.exception;
 
 import lombok.Getter;
+import org.runnect.server.common.constant.ErrorStatus;
 
 @Getter
 public class BasicException extends RuntimeException {
