+++ src/main/java/org/runnect/server/common/exception/NotFoundException.java
@@ -1,5 +1,7 @@
 package org.runnect.server.common.exception;
 
+import org.runnect.server.common.constant.ErrorStatus;
+
 public class NotFoundException extends BasicException {
     public NotFoundException(ErrorStatus errorStatus, String message) {
         super(errorStatus, message);
