+++ src/main/java/org/runnect/server/common/exception/PermissionDeniedException.java
@@ -1,5 +1,7 @@
 package org.runnect.server.common.exception;
 
+import org.runnect.server.common.constant.ErrorStatus;
+
 public class PermissionDeniedException extends BasicException {
 
     public PermissionDeniedException(ErrorStatus errorStatus, String message) {
