+++ src/main/java/org/runnect/server/common/exception/UnauthorizedException.java
@@ -1,5 +1,7 @@
 package org.runnect.server.common.exception;
 
+import org.runnect.server.common.constant.ErrorStatus;
+
 public class UnauthorizedException extends BasicException {
 
     public UnauthorizedException(ErrorStatus errorStatus, String message) {
