+++ src/main/java/org/runnect/server/config/jwt/JwtService.java
@@ -12,8 +12,8 @@ import java.util.Base64;
 import java.util.Date;
 import javax.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
-import org.runnect.server.common.exception.BasicException;
-import org.runnect.server.common.exception.ErrorStatus;
+import org.runnect.server.common.constant.TokenStatus;
+import org.runnect.server.config.redis.RedisService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
