+++ src/main/java/org/runnect/server/course/controller/CourseController.java
@@ -4,8 +4,8 @@ import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.runnect.server.common.dto.ApiResponseDto;
 import org.runnect.server.common.exception.BadRequestException;
-import org.runnect.server.common.exception.ErrorStatus;
-import org.runnect.server.common.exception.SuccessStatus;
+import org.runnect.server.common.constant.ErrorStatus;
+import org.runnect.server.common.constant.SuccessStatus;
 import org.runnect.server.course.dto.request.CourseCreateRequestDto;
 import org.runnect.server.course.dto.request.UpdateCourseRequestDto;
 import org.runnect.server.course.dto.response.CourseCreateResponseDto;
