+++ src/main/java/org/runnect/server/user/exception/userException/DuplicateNicknameException.java
@@ -1,7 +1,7 @@
 package org.runnect.server.user.exception.userException;
 
 import org.runnect.server.common.exception.BasicException;
-import org.runnect.server.common.exception.ErrorStatus;
+import org.runnect.server.common.constant.ErrorStatus;
 
 public class DuplicateNicknameException extends BasicException {
