+++ src/main/java/org/runnect/server/user/exception/userException/NotFoundUserException.java
@@ -2,7 +2,7 @@ package org.runnect.server.user.exception.userException;
 
 import lombok.Getter;
 import org.runnect.server.common.exception.BasicException;
-import org.runnect.server.common.exception.ErrorStatus;
+import org.runnect.server.common.constant.ErrorStatus;
 
 @Getter
 public class NotFoundUserException extends BasicException {
